Correct UINT32 for non-msvc compilers
Other small changes might be needed too, not sure yet

+++ libs/open303-code/Source/DSPCode/GlobalDefinitions.h
@@ -48,7 +48,7 @@ typedef signed long long INT64;
 #ifdef _MSC_VER
 typedef unsigned __int32 UINT32;
 #else
-typedef unsigned long UINT32;
+typedef unsigned int UINT32;
 #endif
 
 // ...constants for numerical precision issues, denorm, etc.:
